[REFACTOR] Use send_generic_request (#2263)
* Refactor all api calls to it's own module, add logging, fix logging levels

* Proper debug leveling

* Fix last standing issues

* Fix km in logs and unaligned block

* Fix comparison

* No more is
